Refactor blog posts into individual files and update sitemap generation
